Adjust action for using GitHub Environments.
This commit removes the functions that drilled into the directory structure
to determine which files changed, as that approach isn't really compatible
with using GitHub environments. Instead, the action now expects that a
terragrunt.hcl file exists at the root of the repository, and that its
parameters are set via environment variables. This enables a setup where
multiple environments are defined in GitHub, and multiple workflows are set
up for each separate environment, with each workflow only being initiated
if its environment variable files are changed in a PR or a push event.

In addition to that, the AWS account ID and the region are now set via
environment variables as well, simplifying the action calls.